Separate update from user profile view
