+++ android-app/app/src/main/AndroidManifest.xml
@@ -56,6 +56,7 @@
             </intent-filter>
         </activity>
         <activity android:name=".UserProfileActivity" />
+        <activity android:name=".UserProfileUpdateActivity" />
         <!-- End Facebook SDK -->
 
         <provider
