+++ android-app/app/src/main/java/tallerii/stories/UserProfileActivity.java
@@ -1,116 +1,24 @@
 package tallerii.stories;
 
-import android.app.ProgressDialog;
 import android.content.Context;
-import android.content.Intent;
-import android.graphics.Bitmap;
-import android.net.Uri;
 import android.os.Bundle;
-import android.provider.MediaStore;
-import android.support.annotation.NonNull;
-import android.view.View;
-import android.widget.ImageView;
 import android.widget.TextView;
-import android.widget.Toast;
-
-import com.bumptech.glide.Glide;
-import com.firebase.ui.storage.images.FirebaseImageLoader;
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.storage.FirebaseStorage;
-import com.google.firebase.storage.OnProgressListener;
-import com.google.firebase.storage.StorageReference;
-import com.google.firebase.storage.UploadTask;
-import com.google.gson.Gson;
-
-import java.io.IOException;
-import java.util.UUID;
 
 import tallerii.stories.controller.ProfileController;
 import tallerii.stories.network.apimodels.ApplicationProfile;
 
-public class UserProfileActivity extends StoriesAppActivity {
+public class UserProfileActivity extends ProfileActivity {
 
-    public static final String PROFILE_OBJECT = "profile";
-    public static final String PROFILE_ID = "profileId";
-    private ImageView imageView;
-    private Uri filePath;
-    private final int PICK_IMAGE_REQUEST = 71;
-    StorageReference storageReference;
-    ProfileController controller = new ProfileController(this);
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
+    protected void initWithChildResource() {
         setContentView(R.layout.activity_user_profile);
-
-        FirebaseStorage storage = FirebaseStorage.getInstance();
-        storageReference = storage.getReference();
         imageView = findViewById(R.id.profile_picture);
-
-        Bundle bundle = getIntent().getExtras();
-        if (bundle != null && bundle.get(PROFILE_OBJECT) != null) {
-            ApplicationProfile profile = new Gson().fromJson(bundle.getString(PROFILE_OBJECT), ApplicationProfile.class);
-            initializeProfile(profile);
-        } else if (bundle != null && bundle.get(PROFILE_ID) != null) {
-            controller.getUser(bundle.getString(PROFILE_ID));
-        }
-    }
-
-    public void chooseImage(View view) {
-        Intent intent = new Intent();
-        intent.setType("image/*");
-        intent.setAction(Intent.ACTION_GET_CONTENT);
-        startActivityForResult(Intent.createChooser(intent, "Select Picture"), PICK_IMAGE_REQUEST);
     }
 
     @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        if(requestCode == PICK_IMAGE_REQUEST && resultCode == RESULT_OK
-                && data != null && data.getData() != null ) {
-            filePath = data.getData();
-            try {
-                Bitmap bitmap = MediaStore.Images.Media.getBitmap(getContentResolver(), filePath);
-                imageView.setImageBitmap(bitmap);
-            }
-            catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-    }
-
-    public void uploadImage(View view) {
-        if(filePath != null) {
-            final ProgressDialog progressDialog = new ProgressDialog(this);
-            progressDialog.setTitle("Uploading...");
-            progressDialog.show();
-
-            StorageReference ref = storageReference.child("images/" + UUID.randomUUID());
-            ref.putFile(filePath)
-                    .addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>() {
-                        @Override
-                        public void onSuccess(UploadTask.TaskSnapshot taskSnapshot) {
-                            progressDialog.dismiss();
-                            Toast.makeText(UserProfileActivity.this, "Uploaded", Toast.LENGTH_SHORT).show();
-                        }
-                    })
-                    .addOnFailureListener(new OnFailureListener() {
-                        @Override
-                        public void onFailure(@NonNull Exception e) {
-                            progressDialog.dismiss();
-                            Toast.makeText(UserProfileActivity.this, "Failed "+e.getMessage(), Toast.LENGTH_SHORT).show();
-                        }
-                    })
-                    .addOnProgressListener(new OnProgressListener<UploadTask.TaskSnapshot>() {
-                        @Override
-                        public void onProgress(UploadTask.TaskSnapshot taskSnapshot) {
-                            double progress = (100.0*taskSnapshot.getBytesTransferred()/taskSnapshot
-                                    .getTotalByteCount());
-                            progressDialog.setMessage("Uploaded "+(int)progress+"%");
-                        }
-                    });
-        }
+    protected ProfileController getController() {
+        return new ProfileController(this);
     }
 
     @Override
@@ -118,18 +26,8 @@ public class UserProfileActivity extends StoriesAppActivity {
         return UserProfileActivity.this;
     }
 
-    public void initializeProfile(ApplicationProfile applicationProfile) {
+    protected void setUserName(ApplicationProfile applicationProfile) {
         TextView username = findViewById(R.id.user_name);
         username.setText(applicationProfile.getFullName());
-
-        TextView friendsCount = findViewById(R.id.friend_count);
-        friendsCount.setText(String.valueOf(applicationProfile.getFriends().size()));
-
-        //try to obtain profile pic from Firebase
-        Glide.with(this /* context */)
-                .using(new FirebaseImageLoader())
-                .load(storageReference.child("images/" + applicationProfile.getProfilePicture()))
-                .into(imageView)
-        ;
     }
 }
